Feature/apply job

+++ backend/job/views.py
@@ -1,12 +1,14 @@
 from django.shortcuts import get_object_or_404
 from django.db.models import Avg, Min, Max, Count
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
+from django.utils import timezone
 from rest_framework import status
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.pagination import PageNumberPagination
 
-from .serializers import JobSerializer
-from .models import Job
+from .serializers import CandiatesAppliedSerializer, JobSerializer
+from .models import CandidatesApplied, Job
 from .filters import JobsFilter
 
 # Create your views here.
@@ -34,12 +36,15 @@ def getAllJobs(request):
 @api_view(['GET'])
 def getJobById(request, id):
     job = get_object_or_404(Job, id=id)
+    candidates = job.candidatesapplied_set.all().count()
     serializer = JobSerializer(job, many=False)
-    return Response(serializer.data)
+    return Response({'job': serializer.data, 'candidates': candidates})
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def createJob(request):
+    request.data['user'] = request.user
     data = request.data
     job = Job.objects.create(**data)
     serializer = JobSerializer(job, many=False)
@@ -47,8 +52,14 @@ def createJob(request):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, id):
     job = get_object_or_404(Job, id=id)
+    if job.user != request.user:
+        return Response(
+            {'message': 'You can not update this job'},
+            status=status.HTTP_403_FORBIDDEN
+        )
     job.title = request.data['title']
     job.description = request.data['description']
     job.email = request.data['email']
@@ -71,8 +82,14 @@ def updateJob(request, id):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, id):
     job = get_object_or_404(Job, id=id)
+    if job.user != request.user:
+        return Response(
+            {'message': 'You can not delete this job'},
+            status=status.HTTP_403_FORBIDDEN
+        )
     job.delete()
     return Response({'message': 'Job deleted.'}, status=status.HTTP_200_OK)
 
@@ -97,3 +114,77 @@ def getTopicStats(request, topic):
         max_salary=Max('salary'),
     )
     return Response(stats)
+
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def applyToJob(request, jobId):
+    user = request.user
+    job = get_object_or_404(Job, id=jobId)
+    if user.userprofile.resume == '':
+        return Response(
+            {'error': 'Please upload your resume first.'},
+            status=status.HTTP_400_BAD_REQUEST
+        )
+    if job.lastDate < timezone.now():
+        return Response(
+            {'error': 'You can not apply to this job, date is over.'},
+            status=status.HTTP_400_BAD_REQUEST
+        )
+    alreadyApplied = job.candidatesapplied_set.filter(user=user).exists()
+    if alreadyApplied:
+        return Response(
+            {'error': 'You have already applied to this job.'},
+            status=status.HTTP_400_BAD_REQUEST
+        )
+    jobApplied = CandidatesApplied.objects.create(
+        job=job,
+        user=user,
+        resume=user.userprofile.resume
+    )
+    return Response(
+        {'applied': True, 'job_id': jobApplied.id},
+        status=status.HTTP_200_OK
+    )
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCurrentUserAppliedJobs(request):
+    args = {'user_id': request.user.id}
+    jobs = CandidatesApplied.objects.filter(**args)
+    serializer = CandiatesAppliedSerializer(jobs, many=True)
+    return Response(serializer.data)
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def isApplied(request, jobId):
+    user = request.user
+    job = get_object_or_404(Job, id=jobId)
+    applied = job.candidatesapplied_set.filter(user=user).exists()
+    return Response(applied)
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCurrentUserJobs(request):
+    args = {'user': request.user.id}
+    jobs = Job.objects.filter(**args)
+    serializer = JobSerializer(jobs, many=True)
+    return Response(serializer.data)
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCandidatesApplied(request, jobId):
+    user = request.user
+    job = get_object_or_404(Job, id=jobId)
+    if job.user != user:
+        return Response(
+            {'error': 'You can not access this job.'},
+            status=status.HTTP_403_FORBIDDEN
+        )
+    candidates = job.candidatesapplied_set.all()
+    serializer = CandiatesAppliedSerializer(candidates, many=True)
+    return Response(serializer.data)
